Clean up plot types
